fix: tweaking the testbed to support what we wanna be doing
