+++ edx_arch_experiments/tests/settings.py
@@ -1,0 +1,18 @@
+import os
+from edx_arch_experiments.settings import *  # noqa
+
+INSTALLED_APPS = [
+    'edx_arch_experiments',
+    'edx_arch_experiments.tests.test_management',
+]
+
+DATABASES = {
+    'default': {
+        'ENGINE': os.environ.get('DB_ENGINE', 'django.db.backends.sqlite3'),
+        'NAME': os.environ.get('DB_NAME', 'default.db'),
+        'USER': os.environ.get('DB_USER', ''),
+        'PASSWORD': os.environ.get('DB_PASSWORD', ''),
+        'HOST': os.environ.get('DB_HOST', ''),
+        'PORT': os.environ.get('DB_PORT', ''),
+    },
+}
