+++ tox.ini
@@ -31,7 +31,7 @@ match-dir = (?!migrations)
 
 [pytest]
 DJANGO_SETTINGS_MODULE = test_settings
-addopts = --cov edx_arch_experiments --cov-report term-missing --cov-report xml
+addopts = --cov edx_arch_experiments --cov-report term-missing --cov-report xml --ds=edx_arch_experiments.tests.settings
 norecursedirs = .* requirements site-packages
 
 [testenv]
